CON-2379: splitted contents of the changeHistory actions module - created separate modules - updated local imports - cleaned up the view reducer
