--- actions/history/changeHistory.js
@@ -1,93 +1,0 @@
-/**
- * Copyright (c) 2017, Shopgate, Inc. All rights reserved.
- *
- * This source code is licensed under the Apache 2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import curry from 'lodash/curry';
-import { logger } from '@shopgate/pwa-core/helpers';
-import { history, parseObjectToQueryString } from '../../helpers/router';
-import { isFunction, isObject } from '../../helpers/validation';
-
-const GO_METHOD = 'go';
-const PUSH_METHOD = 'push';
-const REPLACE_METHOD = 'replace';
-
-/**
- * A higher order thunk that is used to make changes to the router history.
- * @param {string} method The method to be called on the router history.
- * @param {string|Object} location A location string or object.
- * @see https://www.npmjs.com/package/history#navigation
- * @returns {Function} A redux thunk.
- */
-const changeHistory = (method, location) => () => {
-  if (!isFunction(history[method])) {
-    logger.error(`Unable to call method '${method}' on router history.`);
-  }
-
-  let processedLocation = location;
-
-  if (method === REPLACE_METHOD) {
-    // Merge new location with previous location for convenience.
-    processedLocation = {
-      ...history.location,
-      ...processedLocation,
-    };
-  }
-
-  if (isObject(location) && isObject(location.params)) {
-    // Parse location.params to a search query string for convenience.
-    processedLocation = {
-      ...processedLocation,
-      search: parseObjectToQueryString(location.params),
-    };
-  }
-
-  history[method](processedLocation);
-};
-
-const makeHistoryChange = curry(changeHistory);
-
-export const goHistory = makeHistoryChange(GO_METHOD);
-
-/**
- * Goes back in history.
- * @param {number} [amount=1] The optional number of steps to go back. Defaults to 1.
- * @returns {Function} A redux thunk.
- */
-export const goBackHistory = (amount = 1) => () => {
-  if (amount > 0) {
-    history.go(-amount);
-  }
-};
-
-/**
- * Pushes a new location on top of the history.
- * @param {Object|string} location The location to push to the history.
- * @returns {Function} A redux thunk.
- */
-export const pushHistory = makeHistoryChange(PUSH_METHOD);
-
-/**
- * Replaces the current location in the history.
- * @param {Object|string} location The location to set.
- * @returns {Function} A redux thunk.
- */
-export const replaceHistory = makeHistoryChange(REPLACE_METHOD);
-
-/**
- * Resets the history, clearing any previous entry.
- * @param {Object|string|null} [location=null] The location to set after the reset.
- * @returns {Function} A redux thunk.
- */
-export const resetHistory = (location = null) => (dispatch, getState) => {
-  const numEntries = getState().history.length;
-  // Go back the exact amount of entries stored in the history.
-  dispatch(goBackHistory(numEntries - 1));
-
-  // Set the new location if requested.
-  if (location) {
-    dispatch(pushHistory(location));
-  }
-};
